Attributes dialog update: * Move all logic for topmost TreeView in `attributes_dialog.py` to `generic_attributes_treeview.py` * Use `generic_properties_treeview.py` for `additional_tree_frame` if `daq.IDevice.can_cast_from(node)` allowing for modifications * Introduce `data_descriptor_treeview.py` and use it for `additional_tree_frame` if `daq.ISignal.can_cast_from(node)` (it's read only) * Set parent for `simpledialog.ask...` to prevent pop-ups under current window
